requirements.txt
click
Jinja2
MarkupSafe
numpy
pex
untangle
black
flask
dropbox
pyyaml
redis
aiohttp
msgpack
zstandard
